feat: support inputs without version prefix
Previously, passing an input without a 'v' prefix (ex: 1.2.3 instead of
v1.2.3) would fail due to limitations of the golang semver library. This
is especially annoying as this type of versioning seems to be standard
in node projects (package.json), leaving the user to do some
string-manip before passing to semver.

This change allows these types of inputs to be supported by checking
whether the prefix exists and conditionally trimming the final output to
match the user's input.